+++ src/components/editor/editor.js
@@ -489,9 +489,13 @@ Editor.prototype = {
           this.settings.fontpickerSettings.popupmenuSettings = { attributes: fpAttributes };
         }
       }
-
       fpElement.fontpicker(this.settings.fontpickerSettings);
       this.fontPickerElem = fpElement;
+      this.fontPickerElem.tooltip({
+        placementOpts: {
+          x: 40
+        }
+      });
     }
 
     // Invoke Colorpicker, if applicable
@@ -1020,7 +1024,7 @@ Editor.prototype = {
 
       unorderedlist: `<button type="button" class="btn btn-editor" title="${Locale.translate('UnorderedList')}" data-action="insertunorderedlist" data-element="ul">${buttonLabels.unorderedlist}</button>`,
 
-      fontPicker: `<button type="button" class="btn btn-editor fontpicker" data-action="fontStyle" data-init="false"><span>${'FontPicker'}</span></button>`,
+      fontPicker: `<button type="button" class="btn btn-editor fontpicker" title="${Locale.translate('ChooseFont')}" data-action="fontStyle" data-init="false"><span>${'FontPicker'}</span></button>`,
 
       justifyLeft: `<button type="button" class="btn btn-editor" title="${Locale.translate('JustifyLeft')}" data-action="justifyLeft" >${buttonLabels.justifyLeft}</button>`,
 
@@ -1589,7 +1593,7 @@ Editor.prototype = {
   },
 
   /**
-   * Check and set the active states on toolba rbuttons.
+   * Check and set the active states on toolbar buttons.
    * @private
    */
   checkActiveButtons() {