+++ docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # What's New with Enterprise
 
+## v4.55.0 Fixes
+
+- `[Charts]` Fixed a bug where automation ids is not properly rendered on legend, text and slices. ([#5441](https://github.com/infor-design/enterprise/issues/5441))
+
 ## v4.54.0 Features
 
 - `[ApplicationMenu]` - Added the ability to resize the app menu. ([#5193](https://github.com/infor-design/enterprise/issues/5193))
@@ -30,8 +34,11 @@
 - `[Datagrid]` Fixed a bug that prevented the headers of the right frozen columns as well as the order date column from being exported properly. ([#5332](https://github.com/infor-design/enterprise/issues/5332))
 - `[Datagrid]` Fixed a bug where the font color on tags was black when a row was hovered over in dark mode. Font color now white. ([#5289](https://github.com/infor-design/enterprise/issues/5289))
 - `[Datagrid]` Fixed issues with NaN displaying on Decimal and Dropdown inputs when blank options are selected. ([#5395](https://github.com/infor-design/enterprise/issues/5395))
+- `[Datagrid]` Fixed a bug where filter options were unable to reopen after doing pagination and clicking other filter options. ([#5286](https://github.com/infor-design/enterprise/issues/5286))
+- `[Datepicker]` Fixed a bug where the -/+ keys were not detected in datepicker. ([#5353](https://github.com/infor-design/enterprise/issues/5353))
 - `[Donut]` Changed legend design when item exceeds maximum width of chart. ([#5292](https://github.com/infor-design/enterprise/issues/5292))
 - `[Dropdown]` Fixed a bug where backspace in Dropdown is not working when pressed. ([#5113](https://github.com/infor-design/enterprise/issues/5113))
+- `[Editor]` Added tooltip in fontpicker. ([#5472](https://github.com/infor-design/enterprise/issues/5472))
 - `[Fileupload]` Fixed a bug where the required asterisk does not appear on the labels associated with required fields. ([#5285](https://github.com/infor-design/enterprise/issues/5285))
 - `[Homepage]` Adjusted height and width of example homepage ([#5425](https://github.com/infor-design/enterprise/issues/5425))
 - `[Icon]` Changed button icon colors to slate6 ([#5307](https://github.com/infor-design/enterprise/issues/5307))